메인 만다라트 성취도 수정

+++ app/src/main/kotlin/com/emirim/hyejin/mokgongso/model/GetMandal.kt
@@ -4,6 +4,6 @@ import com.google.gson.annotations.SerializedName
 
 data class GetMandal (
         @SerializedName("title")var title: String,
-        @SerializedName("achievement")var achievement: Int,
+        @SerializedName("achievement")var achievement: Double,
         @SerializedName("mandal")var mandal: Array<MandalGet>
 )
